# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Planificacion Horaria",
        "description": "Gestion de horarios de clases",
        "contact": {
            "email": "contacto@tudominio.com"
        },
        "license": {
            "name": "Licencia",
            "url": "https://www.terciarioUrquiza.com.ar/planificacionHoraria"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/docentes": {
            "get": {
                "tags": [
                    "Docente"
                ],
                "summary": "Obtener todos los docentes",
                "description": "Devuelve todos los docentes",
                "operationId": "obtenerDocente",
                "responses": {
                    "200": {
                        "description": "Docentes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Docente"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener los docentes"
                    }
                }
            }
        },
        "/api/docentes/{id}": {
            "get": {
                "tags": [
                    "Docente"
                ],
                "summary": "Obtener docente por dni",
                "description": "Devuelve un docente",
                "operationId": "obtenerDocentePorDni",
                "parameters": [
                    {
                        "name": "dni",
                        "in": "path",
                        "description": "DNI del docente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Docente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Docente"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Docente no encontrado"
                    },
                    "500": {
                        "description": "Error al obtener el docente"
                    }
                }
            }
        },
        "/api/docentes/guardar": {
            "post": {
                "tags": [
                    "Docente"
                ],
                "summary": "Guardar docente",
                "description": "Guarda un docente",
                "operationId": "guardarDocente",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Docente"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Docente guardado correctamente"
                    },
                    "500": {
                        "description": "Error al guardar el docente"
                    }
                }
            }
        },
        "/api/docentes/actualizar/{id}": {
            "put": {
                "tags": [
                    "Docente"
                ],
                "summary": "Actualizar docente",
                "description": "Actualiza un docente",
                "operationId": "actualizarDocente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del docente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Docente"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Docente actualizado correctamente"
                    },
                    "404": {
                        "description": "Docente no encontrado"
                    },
                    "500": {
                        "description": "Error al actualizar el docente"
                    }
                }
            }
        },
        "/api/docentes/eliminar/{id}": {
            "delete": {
                "tags": [
                    "Docente"
                ],
                "summary": "Eliminar docente",
                "description": "Elimina un docente",
                "operationId": "eliminarDocente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del docente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Docente eliminado correctamente"
                    },
                    "404": {
                        "description": "Docente no encontrado"
                    },
                    "500": {
                        "description": "Error al eliminar el docente"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Docente": {
                "title": "Docente",
                "description": "Esquema del objeto Docente",
                "properties": {
                    "dni": {
                        "description": "DNI del docente",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "Nombre del docente",
                        "type": "string"
                    },
                    "apellido": {
                        "description": "Apellido del docente",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email del docente",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}